refactor: ♻️ refactor resize and drag container
