--- apps/builder/src/page/App/components/ScaleSquare/components/InnerResizingContainer/interface.ts
@@ -1,14 +1,0 @@
-import { ReactNode } from "react"
-
-export interface ResizingContainerProps {
-  width: number
-  height: number
-  minWidth: number
-  minHeight: number
-  displayName: string
-  children: ReactNode
-  widgetType: string
-  columnNumber: number
-  widgetTop: number
-  parentNodeDisplayName: string
-}
