+++ apps/builder/src/page/App/components/ScaleSquare/components/ResizingAndDragContainer/ResizeHandler/index.tsx
