+++ apps/builder/src/page/App/components/ScaleSquare/components/ResizingAndDragContainer/ResizeHandler/style.ts
