+++ apps/builder/src/page/App/components/ScaleSquare/components/ResizingAndDragContainer/style.ts
