+++ apps/builder/src/page/App/components/ScaleSquare/components/ResizingAndDragContainer/utils.tsx
