+++ apps/builder/src/page/App/components/ScaleSquare/components/WrapperContainer/interface.ts
@@ -4,6 +4,5 @@ export interface WrapperContainerProps {
   displayName: string
   parentNodeDisplayName: string
   widgetHeight: number
-  widgetType: string
   children: ReactNode
 }
