+++ apps/builder/src/page/App/components/ScaleSquare/utils/useScaleStateSelector.ts
@@ -2,7 +2,7 @@ import { useMemo } from "react"
 import { useSelector } from "react-redux"
 import { getIsILLAEditMode } from "@/redux/config/configSelector"
 import { getExecutionError } from "@/redux/currentApp/executionTree/executionSelector"
-import { getScaleSquareState } from "../components/ResizingContainer/utils"
+import { getScaleSquareState } from "../components/ResizingAndDragContainer/utils"
 
 export const useScaleStateSelector = (displayName: string) => {
   const errors = useSelector(getExecutionError)
